# cmd/task-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Масштабируемая микросервисная CRM-система на Go",
        "title": "CRM Distributed API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Торощина Екатерина",
            "email": "katya@crm.ru"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/federations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт верхний уровень иерархии организации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "federations"
                ],
                "summary": "Создать федерацию",
                "parameters": [
                    {
                        "description": "Название федерации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/crm-distributed_shared_domain.Federation"
                        }
                    }
                }
            }
        },
        "/federations/{uuid}/companies": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт компанию внутри федерации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "federations"
                ],
                "summary": "Создать компанию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID федерации",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Название компании",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "crm-distributed_shared_domain.Company": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "createdByUUID": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "federationUUID": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/crm-distributed_shared_domain.CompanyField"
                    }
                },
                "meta": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/crm-distributed_shared_domain.Project"
                    }
                },
                "projectsTotal": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userTotal": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/crm-distributed_shared_domain.CompanyUser"
                    }
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.CompanyField": {
            "type": "object",
            "required": [
                "companyUUID"
            ],
            "properties": {
                "companyUUID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "dataType": {
                    "maximum": 14,
                    "minimum": 0,
                    "allOf": [
                        {
                            "$ref": "#/definitions/crm-distributed_shared_domain.FieldDataType"
                        }
                    ]
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 5000
                },
                "hash": {
                    "type": "string"
                },
                "icon": {
                    "type": "string",
                    "maxLength": 50
                },
                "meta": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "projectUUID": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "requiredOnStatuses": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "style": {
                    "type": "string",
                    "maxLength": 20
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.CompanyUser": {
            "type": "object",
            "required": [
                "companyUUID",
                "federationUUID",
                "uuid"
            ],
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "companyUUID": {
                    "type": "string"
                },
                "federationUUID": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.User"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.Federation": {
            "type": "object",
            "properties": {
                "companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/crm-distributed_shared_domain.Company"
                    }
                },
                "companiesTotal": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "description": "email создателя",
                    "type": "string"
                },
                "createdByUUID": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "meta": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/crm-distributed_shared_domain.FederationUser"
                    }
                },
                "usersTotal": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.FederationUser": {
            "type": "object",
            "required": [
                "federationUUID",
                "uuid"
            ],
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "federationUUID": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.User"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.FieldDataType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14
            ],
            "x-enum-varnames": [
                "FieldInteger",
                "FieldFloat",
                "FieldString",
                "FieldText",
                "FieldBool",
                "FieldSwitch",
                "FieldArray",
                "FieldData",
                "FieldDataArray",
                "FieldPhone",
                "FieldLink",
                "FieldEmail",
                "FieldTime",
                "FieldDateTime",
                "FieldPeople"
            ]
        },
        "crm-distributed_shared_domain.ProfilePhotoDTO": {
            "type": "object",
            "properties": {
                "large": {
                    "type": "string"
                },
                "medium": {
                    "type": "string"
                },
                "small": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.ProfilePreferences": {
            "type": "object",
            "properties": {
                "timezone": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.Project": {
            "type": "object",
            "properties": {
                "companyUUID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 5000
                },
                "federationUUID": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/crm-distributed_shared_domain.CompanyField"
                    }
                },
                "fieldsSort": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "meta": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "options": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.ProjectOptions"
                },
                "responsibleBy": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "statusGraph": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.StatusGraph"
                },
                "statusSort": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "statusUpdatedAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/crm-distributed_shared_domain.ProjectUser"
                    }
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.ProjectOptions": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "require_cancelation_comment": {
                    "type": "boolean"
                },
                "require_done_comment": {
                    "type": "boolean"
                },
                "status_enable": {
                    "type": "boolean"
                }
            }
        },
        "crm-distributed_shared_domain.ProjectUser": {
            "type": "object",
            "required": [
                "companyUUID",
                "federationUUID",
                "projectUUID",
                "uuid"
            ],
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "companyUUID": {
                    "type": "string"
                },
                "federationUUID": {
                    "type": "string"
                },
                "projectUUID": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.User"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.StatusGraph": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "string"
                },
                "graph": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "crm-distributed_shared_domain.User": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "hasPhoto": {
                    "type": "boolean"
                },
                "isValid": {
                    "type": "boolean"
                },
                "lname": {
                    "type": "string",
                    "maxLength": 30
                },
                "meta": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 30
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "integer",
                    "format": "int64"
                },
                "photo": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.ProfilePhotoDTO"
                },
                "pname": {
                    "type": "string",
                    "maxLength": 30
                },
                "preferences": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.ProfilePreferences"
                },
                "provider": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "validAt": {
                    "type": "string"
                },
                "validationSendAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}